M487 merge main function of wifi & Eth
