--- vendors/nuvoton/boards/numaker_iot_m487_wifi/aws_demos/application_code/main_eth.c
@@ -1,417 +1,0 @@
-/*
- * Amazon FreeRTOS V1.4.7
- * Copyright (C) 2019 Amazon.com, Inc. or its affiliates.  All Rights Reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
- * the Software, and to permit persons to whom the Software is furnished to do so,
- * subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
- * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
- * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- *
- * http://aws.amazon.com/freertos
- * http://www.FreeRTOS.org
- */
-
-/* Standard includes. */
-#include <time.h>
-
-/* FreeRTOS includes. */
-#include "FreeRTOS.h"
-#include "task.h"
-
-#include "FreeRTOS_IP.h"
-#include "FreeRTOS_Sockets.h"
-
-/* Demo includes */
-#include "aws_demo.h"
-
-/* AWS library includes. */
-#include "iot_system_init.h"
-#include "iot_logging_task.h"
-#include "aws_clientcredential.h"
-#include "aws_application_version.h"
-#include "aws_dev_mode_key_provisioning.h"
-
-
-/* Hardware and starter kit includes. */
-#include "NuMicro.h"
-
-/* Declare the firmware version structure for all to see. */
-#if 0
-const AppVersion32_t xAppFirmwareVersion
-#if defined(__CC_ARM)
-#pragma anon_unions
-#endif
-	= {
-	 .u.x.ucMajor = APP_VERSION_MAJOR,
-   .u.x.ucMinor = APP_VERSION_MINOR,
-   .u.x.usBuild = APP_VERSION_BUILD,
-};
-#endif
-
-/* Sleep on this platform */
-#define Sleep( nMs )    vTaskDelay( pdMS_TO_TICKS( nMs ) );
-
-#define mainDEVICE_NICK_NAME                "Nuvoton_Demo"
-
-#define mainLOGGING_TASK_STACK_SIZE         ( configMINIMAL_STACK_SIZE * 5 )
-#define mainLOGGING_MESSAGE_QUEUE_LENGTH    ( 25 )
-
-
-/* The default IP and MAC address used by the demo.  The address configuration
- * defined here will be used if ipconfigUSE_DHCP is 0, or if ipconfigUSE_DHCP is
- * 1 but a DHCP server could not be contacted.  See the online documentation for
- * more information.  In both cases the node can be discovered using
- * "ping RTOSDemo". */
-static const uint8_t ucIPAddress[ 4 ] =
-{
-    configIP_ADDR0,
-    configIP_ADDR1,
-    configIP_ADDR2,
-    configIP_ADDR3
-};
-static const uint8_t ucNetMask[ 4 ] =
-{
-    configNET_MASK0,
-    configNET_MASK1,
-    configNET_MASK2,
-    configNET_MASK3
-};
-static const uint8_t ucGatewayAddress[ 4 ] =
-{
-    configGATEWAY_ADDR0,
-    configGATEWAY_ADDR1,
-    configGATEWAY_ADDR2,
-    configGATEWAY_ADDR3
-};
-static const uint8_t ucDNSServerAddress[ 4 ] =
-{
-    configDNS_SERVER_ADDR0,
-    configDNS_SERVER_ADDR1,
-    configDNS_SERVER_ADDR2,
-    configDNS_SERVER_ADDR3
-};
-
-/* Default MAC address configuration.  The demo creates a virtual network
- * connection that uses this MAC address by accessing the raw Ethernet data
- * to and from a real network connection on the host PC.  See the
- * configNETWORK_INTERFACE_TO_USE definition for information on how to configure
- * the real network connection to use. */
-const uint8_t ucMACAddress[ 6 ] =
-{
-    configMAC_ADDR0,
-    configMAC_ADDR1,
-    configMAC_ADDR2,
-    configMAC_ADDR3,
-    configMAC_ADDR4,
-    configMAC_ADDR5
-};
-
-/**
- * @brief Application task startup hook.
- */
-void vApplicationDaemonTaskStartupHook( void );
-
-
-/**
- * @brief Initializes the board.
- */
-static void prvMiscInitialization( void );
-
-/*-----------------------------------------------------------*/
-/**
- * @Set up the hardware ready to run this demo.
- */
-static void prvSetupHardware( void );
-/*-----------------------------------------------------------*/
-
-
-/**
- * @brief Application runtime entry point.
- */
-#define mainCHECK_TASK_PRIORITY             ( tskIDLE_PRIORITY + 3UL )
-
-#define mainCHECK_TASK_STACK_SIZE           ( configMINIMAL_STACK_SIZE )
-#define mainCHECK_DELAY                     ( ( portTickType ) 5000 / portTICK_RATE_MS )
-static void vCheckTask( void *pvParameters )
-{
-    portTickType xLastExecutionTime;
-
-    xLastExecutionTime = xTaskGetTickCount();
-
-    printf("Check Task is running ...\n");
-
-    for( ;; )
-    {
-        /* Perform this check every mainCHECK_DELAY milliseconds. */
-        vTaskDelayUntil( &xLastExecutionTime, mainCHECK_DELAY );
-				printf("Check delay ...%d\n",xTaskGetTickCount());
-			  Sleep( 1 );
-    }
-}
- 
-int main( void )
-{
-    /* Perform any hardware initialization that does not require the RTOS to be
-     * running.  */
-    prvMiscInitialization();
-    configPRINTF( ( "FreeRTOS App Ver:%x\n", xAppFirmwareVersion));    
-    configPRINTF( ( "FreeRTOS_IPInit\n" ) );	
-    xTaskCreate( vCheckTask, "Check", mainCHECK_TASK_STACK_SIZE, NULL, mainCHECK_TASK_PRIORITY, NULL );	
-
-    FreeRTOS_IPInit( ucIPAddress,
-                     ucNetMask,
-                     ucGatewayAddress,
-                     ucDNSServerAddress,
-                     ucMACAddress );
-    
-    /* Start the scheduler.  Initialization that requires the OS to be running,
-     * including the WiFi initialization, is performed in the RTOS daemon task
-     * startup hook. */
-	  configPRINTF( ( "vTaskStartScheduler\n" ) );
-    vTaskStartScheduler();
-
-    return 0;
-}
-/*-----------------------------------------------------------*/
-
-static void prvSetupHardware( void )
-{
-    /* Unlock protected registers */
-    SYS_UnlockReg();
-
-    /* Set XT1_OUT(PF.2) and XT1_IN(PF.3) to input mode */
-    PF->MODE &= ~(GPIO_MODE_MODE2_Msk | GPIO_MODE_MODE3_Msk);
-
-    /* Enable External XTAL (4~24 MHz) */
-    CLK_EnableXtalRC(CLK_PWRCTL_HXTEN_Msk);
-
-    /* Waiting for 12MHz clock ready */
-    CLK_WaitClockReady( CLK_STATUS_HXTSTB_Msk);
-
-    /* Set core clock as PLL_CLOCK from PLL */
-    CLK_SetCoreClock(FREQ_192MHZ);
-
-    /* Set both PCLK0 and PCLK1 as HCLK/2 */
-    CLK->PCLKDIV = CLK_PCLKDIV_PCLK0DIV2 | CLK_PCLKDIV_PCLK1DIV2;
-
-    /* Enable IP clock */
-    CLK_EnableModuleClock(UART0_MODULE);
-    CLK_EnableModuleClock(EMAC_MODULE);
-
-    /* Select IP clock source */
-    CLK_SetModuleClock(UART0_MODULE, CLK_CLKSEL1_UART0SEL_HXT, CLK_CLKDIV0_UART0(1));
-
-    // Configure MDC clock rate to HCLK / (127 + 1) = 1.5 MHz if system is running at 192 MHz
-    CLK_SetModuleClock(EMAC_MODULE, 0, CLK_CLKDIV3_EMAC(127));
-
-    /* Update System Core Clock */
-    /* User can use SystemCoreClockUpdate() to calculate SystemCoreClock. */
-    SystemCoreClockUpdate();
-
-
-
-    /* Set GPB multi-function pins for UART0 RXD and TXD */
-    SYS->GPB_MFPH &= ~(SYS_GPB_MFPH_PB12MFP_Msk | SYS_GPB_MFPH_PB13MFP_Msk);
-    SYS->GPB_MFPH |= (SYS_GPB_MFPH_PB12MFP_UART0_RXD | SYS_GPB_MFPH_PB13MFP_UART0_TXD);
-    // Configure RMII pins
-    SYS->GPA_MFPL |= SYS_GPA_MFPL_PA6MFP_EMAC_RMII_RXERR | SYS_GPA_MFPL_PA7MFP_EMAC_RMII_CRSDV;
-    SYS->GPC_MFPL |= SYS_GPC_MFPL_PC6MFP_EMAC_RMII_RXD1 | SYS_GPC_MFPL_PC7MFP_EMAC_RMII_RXD0;
-    SYS->GPC_MFPH |= SYS_GPC_MFPH_PC8MFP_EMAC_RMII_REFCLK;
-    SYS->GPE_MFPH |= SYS_GPE_MFPH_PE8MFP_EMAC_RMII_MDC |
-                     SYS_GPE_MFPH_PE9MFP_EMAC_RMII_MDIO |
-                     SYS_GPE_MFPH_PE10MFP_EMAC_RMII_TXD0 |
-                     SYS_GPE_MFPH_PE11MFP_EMAC_RMII_TXD1 |
-                     SYS_GPE_MFPH_PE12MFP_EMAC_RMII_TXEN;
-
-    // Enable high slew rate on all RMII TX output pins
-    PE->SLEWCTL = (GPIO_SLEWCTL_HIGH << GPIO_SLEWCTL_HSREN10_Pos) |
-                  (GPIO_SLEWCTL_HIGH << GPIO_SLEWCTL_HSREN11_Pos) |
-                  (GPIO_SLEWCTL_HIGH << GPIO_SLEWCTL_HSREN12_Pos);
-
-    /* Lock protected registers */
-    SYS_LockReg();
-
-    /* Init UART to 115200-8n1 for print message */
-    UART_Open(UART0, 115200);
-}
-
-/*-----------------------------------------------------------*/
-static void prvMiscInitialization( void )
-{
-    /* Start logging task. */
-    xLoggingTaskInitialize( mainLOGGING_TASK_STACK_SIZE,
-                            tskIDLE_PRIORITY,
-                            mainLOGGING_MESSAGE_QUEUE_LENGTH );
-
-    prvSetupHardware();
-}
-
-void vApplicationDaemonTaskStartupHook( void )
-{
-  ;
-}
-/*-----------------------------------------------------------*/
-
-void vApplicationIPNetworkEventHook( eIPCallbackEvent_t eNetworkEvent )
-{
-    uint32_t ulIPAddress, ulNetMask, ulGatewayAddress, ulDNSServerAddress;
-    char cBuffer[ 16 ];
-    static BaseType_t xTasksAlreadyCreated = pdFALSE;
-
-    /* If the network has just come up...*/
-    if( eNetworkEvent == eNetworkUp )
-    {
-        /* The network is up so we can run. */
-			  
-        if( ( SYSTEM_Init() == pdPASS ) && ( xTasksAlreadyCreated == pdFALSE ) )
-        {
-          /* A simple example to demonstrate key and certificate provisioning in
-            * microcontroller flash using PKCS#11 interface. This should be replaced
-            * by production ready key provisioning mechanism. */
-            vDevModeKeyProvisioning();
-          
-            /* Run all demos. */
-            DEMO_RUNNER_RunDemos();
-            xTasksAlreadyCreated = pdTRUE;
-        }
-
-        /* Print out the network configuration, which may have come from a DHCP
-         * server. */
-        FreeRTOS_GetAddressConfiguration(
-            &ulIPAddress,
-            &ulNetMask,
-            &ulGatewayAddress,
-            &ulDNSServerAddress );
-        FreeRTOS_inet_ntoa( ulIPAddress, cBuffer );
-        FreeRTOS_printf( ( "\r\n\r\nIP Address: %s\r\n", cBuffer ) );
-
-        FreeRTOS_inet_ntoa( ulNetMask, cBuffer );
-        FreeRTOS_printf( ( "Subnet Mask: %s\r\n", cBuffer ) );
-
-        FreeRTOS_inet_ntoa( ulGatewayAddress, cBuffer );
-        FreeRTOS_printf( ( "Gateway Address: %s\r\n", cBuffer ) );
-
-        FreeRTOS_inet_ntoa( ulDNSServerAddress, cBuffer );
-        FreeRTOS_printf( ( "DNS Server Address: %s\r\n\r\n\r\n", cBuffer ) );
-    }
-
-}
-
-/*-----------------------------------------------------------*/
-
-void vAssertCalled( const char * pcFile,
-                    uint32_t ulLine )
-{
-    const uint32_t ulLongSleep = 1000UL;
-    volatile uint32_t ulBlockVariable = 0UL;
-    volatile char * pcFileName = ( volatile char * ) pcFile;
-    volatile uint32_t ulLineNumber = ulLine;
-
-    ( void ) pcFileName;
-    ( void ) ulLineNumber;
-
-    configPRINTF( ( "vAssertCalled %s, %ld\n", pcFile, ( long ) ulLine ) );
-
-    /* Setting ulBlockVariable to a non-zero value in the debugger will allow
-     * this function to be exited. */
-    taskDISABLE_INTERRUPTS();
-    {
-        while( ulBlockVariable == 0UL )
-        {
-            Sleep( ulLongSleep );
-        }
-    }
-    taskENABLE_INTERRUPTS();
-}
-
-#if ( ipconfigUSE_LLMNR != 0 ) || ( ipconfigUSE_NBNS != 0 ) || ( ipconfigDHCP_REGISTER_HOSTNAME == 1 )
-
-    const char * pcApplicationHostnameHook( void )
-    {
-        /* This function will be called during the DHCP: the machine will be registered 
-         * with an IP address plus this name. */
-        return clientcredentialIOT_THING_NAME;
-    }
-
-#endif
-/*-----------------------------------------------------------*/
-
-#if ( ipconfigUSE_LLMNR != 0 ) || ( ipconfigUSE_NBNS != 0 )
-
-    BaseType_t xApplicationDNSQueryHook( const char * pcName )
-    {
-        BaseType_t xReturn;
-
-        /* Determine if a name lookup is for this node.  Two names are given
-         * to this node: that returned by pcApplicationHostnameHook() and that set
-         * by mainDEVICE_NICK_NAME. */
-        if( strcmp( pcName, pcApplicationHostnameHook() ) == 0 )
-        {
-            xReturn = pdPASS;
-        }
-        else if( strcmp( pcName, mainDEVICE_NICK_NAME ) == 0 )
-        {
-            xReturn = pdPASS;
-        }
-        else
-        {
-            xReturn = pdFAIL;
-        }
-
-        return xReturn;
-    }
-
-#endif /* if ( ipconfigUSE_LLMNR != 0 ) || ( ipconfigUSE_NBNS != 0 ) */
-/*-----------------------------------------------------------*/
-
-
-void vApplicationIdleHook( void )
-{
-    const uint32_t ulMSToSleep = 1;
-    const TickType_t xIdleCheckPeriod = pdMS_TO_TICKS( 1000UL );
-    static TickType_t xTimeNow, xLastTimeCheck = 0;
-
-    /* vApplicationIdleHook() will only be called if configUSE_IDLE_HOOK is set
-     * to 1 in FreeRTOSConfig.h.  It will be called on each iteration of the idle
-     * task.  It is essential that code added to this hook function never attempts
-     * to block in any way (for example, call xQueueReceive() with a block time
-     * specified, or call vTaskDelay()).  If application tasks make use of the
-     * vTaskDelete() API function to delete themselves then it is also important
-     * that vApplicationIdleHook() is permitted to return to its calling function,
-     * because it is the responsibility of the idle task to clean up memory
-     * allocated by the kernel to any task that has since deleted itself. */
-
-    xTimeNow = xTaskGetTickCount();
-
-    if( ( xTimeNow - xLastTimeCheck ) > xIdleCheckPeriod )
-    {
-        /* do something in the idle task */
-        xLastTimeCheck = xTimeNow;
-    }
-
-    /* This is just a trivial example of an idle hook.  It is called on each
-     * cycle of the idle task if configUSE_IDLE_HOOK is set to 1 in
-     * FreeRTOSConfig.h.  It must *NOT* attempt to block.  In this case the
-     * idle task just sleeps to lower the CPU usage. */
-//    Sleep( ulMSToSleep ); /* Call vTaskDelay in IdleHook will cause vTaskSwitchContext failure*/
-
-}
-
-void vApplicationTickHook( void )
-{
-	static uint32_t tickHookCnt=0; 
-	tickHookCnt++;
-}
-/*-----------------------------------------------------------*/
